+++ vshard/storage/CMakeLists.txt
@@ -1,2 +1,2 @@
-install(FILES init.lua reload_evolution.lua ref.lua sched.lua
+install(FILES init.lua reload_evolution.lua ref.lua sched.lua schema.lua
         DESTINATION ${TARANTOOL_INSTALL_LUADIR}/vshard/storage)
